Inherit svelte global context for components
Currently context is created fresh for each component which means that
there is no knowledge of any context set by any parent to the editor.

This is problematic when trying to use context within a svelte
component rendered within a nodeview.

This change inits the context based on the parent but requires the
parent to call this function.

Copies an approach by svelte-preprocess-react where it keeps track of
svelte context and passes it down.